+++ src/composables/use-consent/use-consent.test.ts
@@ -1,13 +1,17 @@
 import flushPromises from "flush-promises";
-import { consent, consentDeniedAll, consentGrantedAll } from "../api/consent";
-import addGtag from "../core/add-gtag";
-import { resetSettings, updateSettings } from "../core/settings";
+import {
+  consent,
+  consentDeniedAll,
+  consentGrantedAll,
+} from "../../api/consent";
+import addGtag from "../../core/add-gtag";
+import { resetSettings, updateSettings } from "../../core/settings";
 import { useConsent } from "./use-consent";
 
-vi.mock("../api/consent");
-vi.mock("../core/add-gtag");
-vi.mock("../core/settings", async () => ({
-  ...(await vi.importActual("../core/settings")),
+vi.mock("../../api/consent");
+vi.mock("../../core/add-gtag");
+vi.mock("../../core/settings", async () => ({
+  ...(await vi.importActual("../../core/settings")),
   updateSettings: vi.fn(),
 }));
