+++ src/composables/use-consent/use-consent.ts
@@ -1,9 +1,13 @@
 import { type Ref, ref } from "vue";
-import { consent, consentDeniedAll, consentGrantedAll } from "../api/consent";
-import addGtag from "../core/add-gtag";
-import { type PluginSettings, updateSettings } from "../core/settings";
-import type { GtagConsentParams } from "../types/gtag";
-import { isServer } from "../utils";
+import {
+  consent,
+  consentDeniedAll,
+  consentGrantedAll,
+} from "../../api/consent";
+import addGtag from "../../core/add-gtag";
+import { type PluginSettings, updateSettings } from "../../core/settings";
+import type { GtagConsentParams } from "../../types/gtag";
+import { isServer } from "../../utils";
 
 export type UseWithConsentReturn = {
   hasConsent: Ref<boolean>;
